+++ playground/Playground.tsx
@@ -4,6 +4,7 @@ import React from 'react'
 import { PlaygroundConnection } from './Sections/Playground.Connection'
 import { PlaygroundWallet } from './Sections/Playground.Wallet'
 import { PlaygroundToken } from './Sections/Playground.Token'
+import { PlaygroundUTXO } from './Sections/Playground.UTXO'
 import { createStackNavigator } from '@react-navigation/stack'
 import { NavigationContainer } from '@react-navigation/native'
 
@@ -24,6 +25,11 @@ export function PlaygroundScreen (): JSX.Element {
       <View style={tailwind('mt-4')}>
         <PlaygroundWallet />
       </View>
+
+      <View style={tailwind('mt-4')}>
+        <PlaygroundUTXO />
+      </View>
+
       <View style={tailwind('mt-4')}>
         <PlaygroundToken />
       </View>
