+++ playground/Sections/Playground.Wallet.tsx
@@ -6,13 +6,15 @@ import { PlaygroundStatus } from '../Playground.Status'
 import { WalletStatus } from '../../store/wallet'
 import { useWalletAPI } from '../../hooks/wallet/WalletAPI'
 import { useDispatch } from 'react-redux'
-import { usePlaygroundRpcClient } from '../../hooks/api/usePlaygroundClient'
 
-export function PlaygroundWallet (): JSX.Element {
+export function PlaygroundWallet (): JSX.Element | null {
   const WalletAPI = useWalletAPI()
-  const rpcClient = usePlaygroundRpcClient()
   const status = WalletAPI.getStatus()
   const dispatch = useDispatch()
+
+  if (WalletAPI.getStatus() === WalletStatus.NO_WALLET) {
+    return null
+  }
 
   return (
     <View>
@@ -48,15 +50,6 @@ export function PlaygroundWallet (): JSX.Element {
         testID='playground_wallet_random'
         title='Setup wallet with a randomly generated mnemonic seed'
         onPress={() => WalletAPI.randomMnemonic(dispatch)}
-      />
-
-      <PlaygroundAction
-        testID='playground_wallet_top_up'
-        title='Top up current wallet with 50 DFI UTXO'
-        onPress={async () => {
-          const address = await WalletAPI.getWallet().get(0).getAddress()
-          await rpcClient.wallet.sendToAddress(address, 50)
-        }}
       />
     </View>
   )
